gui: break out the list settings modal into separate classes
